Fix HNSW zombie vector race condition with robust unit test coverage
This commit resolves a race condition in `HnswIndex::add` where concurrent removals could leave "zombie vectors" (unreachable data) in the index. The fix involves re-verifying the ID mapping ownership after acquiring the index write lock.

To ensure 100% code coverage for this rare race condition, we introduce:
1.  `#[cfg(test)]` static atomics in `src/index/vector/hnsw.rs` (`INJECT_RACE_DELAY`, `HIT_RACE_CONDITION`) to allow internal testing of the race window.
2.  A robust unit test `test_zombie_vectors_race_coverage` *inside* `src/index/vector/hnsw.rs` (in `mod tests`) that explicitly enables `INJECT_RACE_DELAY` (500ms) and uses a polling loop (`yield_now`) to guarantee the race window is hit.

Moving the test from integration tests (`tests/havoc_zombies.rs`) to unit tests (`src/index/vector/hnsw.rs`) ensures reliable coverage reporting and simplifies internal state access. The logic was simplified to a single boolean check to prevent coverage gaps.
